+++ source/html/js/app/ui/settings_menu.js
@@ -143,7 +143,7 @@ if (connections.get_current() !== null) {
 $("#save_endpoint_connection").on("click", () => {
     try {
         // trim the string, normalize the link, remove any trailing slash
-        const endpoint = filterXSS($("#input_endpoint_url").val().trim()).replace(/\/+$/, "");
+        const endpoint = filterXSS($("#input_endpoint_url").val().trim()).replace(/\/$/, "");
         const apiKey = filterXSS($("#input_endpoint_key").val().trim());
         const store_locally = $("#connectionRemember").prop("checked");
         console.log("store locally = " + store_locally);